Fix: Fix incorrect IoU calculation in `metrics.py`
Fixed incorrect IoU calculation in `metrics.py`. The previous implementation was calculating the total IoU for each class, and then dividing by the number of class appearances. This is incorrect.
To fix this issue, a method of calculating the total number of True Positives, False Positives, and False Negatives for each class was implemented. Then, using these values, the IoU for each class is calculated. The final IoU is then calculated by averaging the IoU for each class.

Changed the `ignore_classes` list param to a single `ignore_class` int param in `metrics.py`.

Refactor: Refactoring change in `segformer.py`. Separated the SegFormer model factory function for into two functions:
- `create_segformer_model_for_train` - initializes a SegFormer model for training (with `num_classes` param as the model head)
- `create_segformer_model_for_inference` - initializes a SegFormer model for inference (evaluation and testing)