Refactor: Move logits resize to the loss function
Previously, the teacher and student logits resize was done in the trainer. Furthermore, the logits have been resized to match the size of the target. This may cause some problems with the loss calculation. Thus, the logits resize is now done in the loss function.
`utils` module has been created in which a `resize_outputs` function is implemented.

2 separate resizes are done for the teacher and student logits:
- the student logits are resized to match the size of the target,
- whereas the teacher logits are downsized to match the size of the student's logits.